--- frontend/starkbnb-web/app/components/Navbar.tsx
@@ -1,10 +1,0 @@
-
-import React from 'react'
-import Logo from './ui/Logo'
-
-
-
-  )
-}
-
-export default Navbar
