--- frontend/starkbnb-web/app/components/ui/Logo.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Logo = () => {
-  return (
-    <div>STARKBNB</div>
-  )
-}
-
-export default Logo
