+++ frontend/starkbnb-web/app/globals.css
@@ -3,7 +3,7 @@
 @tailwind utilities;
 
 :root {
-  --background: #bad1ced8;
+  --background: #f2f2f2;
   --foreground: #171717;
 }
 
@@ -18,4 +18,6 @@ body {
   color: var(--foreground);
   background: var(--background);
   font-family: Arial, Helvetica, sans-serif;
+  margin: 0;
+  padding-bottom: 5rem;
 }
