+++ frontend/starkbnb-web/components/ui/Logo.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Logo = () => {
+  return (
+    <div>STARKBNB</div>
+  )
+}
+
+export default Logo
