+++ frontend/starkbnb-web/public/brick-building-with-windows-blue-sky.jpg:Zone.Identifier
@@ -1,0 +1,4 @@
+[ZoneTransfer]
+ZoneId=3
+ReferrerUrl=https://www.freepik.com/free-photo/brick-building-with-windows-blue-sky_9633266.htm
+HostUrl=https://downloadscdn6.freepik.com/246466/1/9.jpg?filename=brick-building-with-windows-blue-sky.jpg&token=exp=1732915647~hmac=4002edb8e6e66173c6d695c2f09c49f3&filename=9.jpg
