+++ frontend/starkbnb-web/public/rental-background-1.jpg:Zone.Identifier
@@ -1,0 +1,4 @@
+[ZoneTransfer]
+ZoneId=3
+ReferrerUrl=https://www.freepik.com/premium-vector/hand-drawn-city-background_177634438.htm
+HostUrl=https://img.freepik.com/premium-vector/hand-drawn-city-background_23-2151435769.jpg?w=900
