+++ frontend/starkbnb-web/public/white-concrete-apartment-buildings.jpg:Zone.Identifier
@@ -1,0 +1,4 @@
+[ZoneTransfer]
+ZoneId=3
+ReferrerUrl=https://www.freepik.com/free-photos-vectors/housing-background
+HostUrl=https://downloadscdn6.freepik.com/250224/1/58.jpg?filename=white-concrete-apartment-buildings.jpg&token=exp=1732915627~hmac=2f2a646a2c64682cfd9d46452a29b7b5&filename=58.jpg
